🎨 Format: OngoingChallenge 컴포넌트 제거
