+++ src/pages/ongoingPage/OngoingPage.module.scss
@@ -1,5 +1,5 @@
-@use '../../../../styles/mixins' as m;
-@use '../../../../styles/variables' as *;
+@use '../../styles/mixins' as m;
+@use '../../styles/variables' as *;
 
 .todayChallenge {
   @include m.flex-column-center(flex-start, center, 6px);
